refactor: transfer relations from user to it's own app
