+++ django/ft_transcendence/settings.py
@@ -38,6 +38,7 @@ INSTALLED_APPS = [
     'chat',
     'pong',
     'user',
+    'relations',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
