+++ django/ft_transcendence/urls.py
@@ -22,5 +22,6 @@ urlpatterns = [
     path('', include('accounts.urls')),
     path('', include('chat.urls')),
     path('', include('user.urls')),
+    path('', include('relations.urls')),
     path('admin/', admin.site.urls),
 ]
