+++ django/user/admin.py
@@ -1,15 +1,18 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
-import user.models as models
+
+from user.models import User
+from relations.models import IsFriendsWith
+from relations.models import IsBlockedBy
 
 
 class IsFriendsWithInline(admin.TabularInline):
-    model = models.IsFriendsWith
+    model = IsFriendsWith
     fk_name = 'user1'
 
 
 class IsBlockedByInline(admin.TabularInline):
-    model = models.IsBlockedBy
+    model = IsBlockedBy
     fk_name = 'blocker'
 
 
@@ -17,7 +20,4 @@ class UserAdm(UserAdmin):
     inlines = [IsFriendsWithInline, IsBlockedByInline]
 
 
-admin.site.register(models.User, UserAdm)
-admin.site.register(models.IsFriendsWith)
-admin.site.register(models.IsBlockedBy)
-admin.site.register(models.FriendInvite)
+admin.site.register(User, UserAdm)
